+++ authentication/app/controllers/refinery/registrations_controller.rb
@@ -38,11 +38,11 @@ module ::Refinery
 
         flash[:message] = "<h2>#{t('welcome', :scope => 'refinery.users.create', :who => @user.username).gsub(/\.$/, '')}.</h2>".html_safe
 
-        site_name_setting = ::Refinery::RefinerySetting.find_or_create_by_name('site_name', :value => "Company Name")
+        site_name_setting = ::Refinery::Setting.find_or_create_by_name('site_name', :value => "Company Name")
         if site_name_setting.value.to_s =~ /^(|Company\ Name)$/ or Role[:refinery].users.count == 1
           flash[:message] << "<p>#{t('setup_website_name_html', :scope => 'refinery.users',
-                                     :link => edit_refinery_admin_refinery_setting_path(site_name_setting, :dialog => true),
-                                     :title => t('edit', :scope => 'refinery.admin.refinery_settings'))}</p>".html_safe
+                                     :link => edit_refinery_admin_setting_path(site_name_setting, :dialog => true),
+                                     :title => t('edit', :scope => 'refinery.admin.settings'))}</p>".html_safe
         end
         sign_in(@user)
         redirect_back_or_default(refinery_admin_root_path)