+++ dashboard/app/controllers/refinery/admin/dashboard_controller.rb
@@ -5,7 +5,7 @@ module ::Refinery
       def index
         @recent_activity = []
 
-        Refinery::Plugins.active.each do |plugin|
+        ::Refinery::Plugins.active.each do |plugin|
           begin
             plugin.activity.each do |activity|
               @recent_activity << activity.class.where(activity.conditions).
@@ -22,13 +22,13 @@ module ::Refinery
 
         @recent_activity = @recent_activity.flatten.compact.sort { |x,y|
           y.updated_at <=> x.updated_at
-        }.first(activity_show_limit=::Refinery::RefinerySetting.find_or_set(:activity_show_limit, 7))
+        }.first(activity_show_limit=::Refinery::Setting.find_or_set(:activity_show_limit, 7))
 
         @recent_inquiries = defined?(Inquiry) ? Inquiry.latest(activity_show_limit) : []
       end
 
       def disable_upgrade_message
-        ::Refinery::RefinerySetting.set(:show_internet_explorer_upgrade_message, {
+        ::Refinery::Setting.set(:show_internet_explorer_upgrade_message, {
           :value => false,
           :scoping => 'refinery'
         })