+++ db/migrate/20100913234710_create_refinerycms_settings_schema.rb
@@ -1,7 +1,7 @@
 class CreateRefinerycmsSettingsSchema < ActiveRecord::Migration
   def self.up
-    unless ::Refinery::RefinerySetting.table_exists?
-      create_table ::Refinery::RefinerySetting.table_name, :force => true do |t|
+    unless ::Refinery::Setting.table_exists?
+      create_table ::Refinery::Setting.table_name, :force => true do |t|
         t.string   "name"
         t.text     "value"
         t.boolean  "destroyable",             :default => true
@@ -12,12 +12,12 @@ class CreateRefinerycmsSettingsSchema < ActiveRecord::Migration
         t.string   "callback_proc_as_string"
       end
 
-      add_index ::Refinery::RefinerySetting.table_name, ["name"], :name => "index_#{::Refinery::RefinerySetting.table_name}_on_name"
+      add_index ::Refinery::Setting.table_name, ["name"], :name => "index_#{::Refinery::Setting.table_name}_on_name"
     end
   end
 
   def self.down
-    [::Refinery::RefinerySetting].reject{|m|
+    [::Refinery::Setting].reject{|m|
       !(defined?(m) and m.respond_to?(:table_name))
     }.each do |model|
       drop_table model.table_name if model.table_exists?