+++ db/migrate/20100926142529_add_value_type_to_refinery_settings.rb
@@ -1,13 +1,13 @@
 class AddValueTypeToRefinerySettings < ActiveRecord::Migration
   def self.up
-    add_column ::Refinery::RefinerySetting.table_name, :form_value_type, :string
+    add_column ::Refinery::Setting.table_name, :form_value_type, :string
 
-    ::Refinery::RefinerySetting.reset_column_information
+    ::Refinery::Setting.reset_column_information
   end
 
   def self.down
-    remove_column ::Refinery::RefinerySetting.table_name, :form_value_type
+    remove_column ::Refinery::Setting.table_name, :form_value_type
 
-    ::Refinery::RefinerySetting.reset_column_information
+    ::Refinery::Setting.reset_column_information
   end
 end