+++ pages/app/models/refinery/page.rb
@@ -65,8 +65,8 @@ module Refinery
     has_friendly_id :title, :use_slug => true,
                     :default_locale => (::Refinery::I18n.default_frontend_locale rescue :en),
                     :reserved_words => %w(index new session login logout users refinery admin images wymiframe),
-                    :approximate_ascii => ::Refinery::RefinerySetting.find_or_set(:approximate_ascii, false, :scoping => "pages"),
-                    :strip_non_ascii => ::Refinery::RefinerySetting.find_or_set(:strip_non_ascii, false, :scoping => "pages")
+                    :approximate_ascii => ::Refinery::Setting.find_or_set(:approximate_ascii, false, :scoping => "pages"),
+                    :strip_non_ascii => ::Refinery::Setting.find_or_set(:strip_non_ascii, false, :scoping => "pages")
 
     has_many :parts,
              :class_name => "::Refinery::PagePart",
@@ -265,7 +265,7 @@ module Refinery
     class << self
       # Accessor to find out the default page parts created for each new page
       def default_parts
-        ::Refinery::RefinerySetting.find_or_set(:default_page_parts, ["Body", "Side Body"])
+        ::Refinery::Setting.find_or_set(:default_page_parts, ["Body", "Side Body"])
       end
 
       # Wraps up all the checks that we need to do to figure out whether
@@ -283,7 +283,7 @@ module Refinery
       end
 
       def use_marketable_urls?
-        ::Refinery::RefinerySetting.find_or_set(:use_marketable_urls, true, :scoping => 'pages')
+        ::Refinery::Setting.find_or_set(:use_marketable_urls, true, :scoping => 'pages')
       end
 
       def expire_page_caching