+++ public/javascripts/translations.js
@@ -1,2 +1,2 @@
 var I18n = I18n || {};
-I18n.translations = {"en":{"js":{"admin":{"confirm_changes":"Any changes you've made will be lost. Are you sure you want to continue without saving?"}}},"fi":{"js":{"admin":{"confirm_changes":"Kaikki muutoksesi menetet\u00e4\u00e4n. Oletko varma, ett\u00e4 haluat jatkaa tallentamatta?"}}}};
+I18n.translations = {"fi":{"js":{"admin":{"confirm_changes":"Kaikki muutoksesi menetet\u00e4\u00e4n. Oletko varma, ett\u00e4 haluat jatkaa tallentamatta?"}}}};