+++ settings/config/routes.rb
@@ -2,7 +2,7 @@
   scope(:path => 'refinery', :as => 'refinery_admin', :module => 'refinery/admin') do
     resources :settings,
               :except => :show,
-              :as => :refinery_settings,
-              :controller => :refinery_settings
+              :as => :settings,
+              :controller => :settings
   end
 end